Replace SERIAL -> BIGSERIAL

+++ server/db setup/tables/permissions.sql
@@ -1,5 +1,5 @@
 CREATE TABLE permissions (
-    id INTEGER PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     title TEXT NOT NULL,
     description TEXT NOT NULL
 )
